[iOS] Rebuild MobileGestalt cache if needed https://bugs.webkit.org/show_bug.cgi?id=211739
Reviewed by Darin Adler.

Source/WebCore/PAL:

Add SPI to check if MobileGestalt cache is valid, and declare MobileGestaltHelperProxy class.

* pal/spi/ios/MobileGestaltSPI.h:

Source/WebKit:

This patch was written by Geoff Garen and myself. Rebuild the MobileGestalt cache in the UI process if the
MobileGestalt cache is stale. This should be done before starting a new WebContent process. If the cache
is not rebuilt, every MobileGestalt query in the WebContent process will fail, since access to the daemon
is denied.

No new tests, since testing this is not trivial. It would involve deleting the MobileGestalt cache and
making sure it is rebuilt afterwards. This could be an API test, but API tests are not supported on iOS
devices.

* UIProcess/Cocoa/WebProcessPoolCocoa.mm:
(WebKit::WebProcessPool::platformInitializeWebProcess):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261584 268f45cc-cd09-0410-ab3c-d52691b4dbfc