Exception check for OOM is a bit too late in JSBigInt::exponentiate. https://bugs.webkit.org/show_bug.cgi?id=211823
Reviewed by Mark Lam.

JSTests:

Just add one more case so that this test now catches this bug (the control-flow through JSBigInt::exponentiate is pretty convoluted).

* stress/bigint-exponential-oom.js:
(shouldThrow):

Source/JavaScriptCore:

We were doing multiplyImpl(...).payload.asHeapBigInt(), but multiplyImpl can return a null payload if it causes an exception.
So we must first check whether an exception was raised, and only if not can we do asHeapBigInt.

* runtime/JSBigInt.cpp:
(JSC::JSBigInt::exponentiateImpl):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261596 268f45cc-cd09-0410-ab3c-d52691b4dbfc