MovHint can see an arguments object be MovHinted to a Tmp https://bugs.webkit.org/show_bug.cgi?id=211820 <rdar://problem/62882158>
Reviewed by Keith Miller.

JSTests:

* stress/varargs-forwarding-can-see-arguments-mov-hint-to-for-of-tmp.js: Added.
(let.iter.Symbol.iterator.return.next):
(let.iter.Symbol.iterator):

Source/JavaScriptCore:

We had an assert that it wasn't possible to have a MovHint from an arguments
object to a Tmp. However, this is possible with for-of. There is nothing
about the current algorithm that is specific to only VirtualRegisters. The
algorithm also works over Tmps. So I've generalized the algorithm to just work
over Operand.

* dfg/DFGVarargsForwardingPhase.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261603 268f45cc-cd09-0410-ab3c-d52691b4dbfc