Web Inspector: Fails to pretty-print a particular CSS file https://bugs.webkit.org/show_bug.cgi?id=211930
Reviewed by Joseph Pecoraro.

Source/WebInspectorUI:

* UserInterface/Workers/Formatter/CSSFormatter.js:
(CSSFormatter.prototype._format):
Keep a stack of special sequences (e.g. `"`, `'`, `/*`, etc.), only outputting the text in
between the start and end of the sequence when the stack is empty. Ignore all other special
sequences when in a comma sequence. Add proper checks for if the star/end is escaped.
Drive-by: minor refactor so that the arrow functions are created outside the loop.

LayoutTests:

* inspector/formatting/formatting-css.html:
* inspector/formatting/formatting-css-expected.txt:
* inspector/formatting/resources/css-tests/url.css: Added.
* inspector/formatting/resources/css-tests/url-expected.css: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261772 268f45cc-cd09-0410-ab3c-d52691b4dbfc