Add iterator checking to ListHashSet https://bugs.webkit.org/show_bug.cgi?id=211669
Reviewed by Anders Carlsson.

Source/WebCore:

* page/ios/ContentChangeObserver.h: Added an include of Element.h, needed to call
makeWeakPtr in an inline function. This is due to a change in the way makeWeakPtr
now checks the type of the argument. It's possible we could refine it further to
relax this requirement, but it seems OK to include Element.h here.

Source/WTF:

HashSet and HashMap have iterator checking in debug builds.
Add similar checking to ListHashSet, controlled by the same
macro, CHECK_HASHTABLE_ITERATORS. Use WeakPtr to make the
implementation simple.

* wtf/Forward.h: Update to add a second parameter to WeakPtr.
Also rearranged and tweaked the file a bug.

* wtf/ListHashSet.h: Make ListHashSet and ListHashSetNode derive
from CanMakeWeakPtr. Add m_weakSet and m_weakPosition members to
ListHashSetConstIterator, and assert their values at the appropriate
times so we will get a breakpoint or crash.

* wtf/WeakHashSet.h: Updated to add a Counter argument for testing.

* wtf/WeakPtr.h: Removed the DID_CREATE_WEAK_PTR_IMPL and
WILL_DESTROY_WEAK_PTR_IMPL macros, which didn't really work because
using them violated the C++ one-definition rule. Replaced with
a Counter argument, which defaults to EmptyCounter, which is inlines
that do nothing. This required some updates to the classes and
functions to make them work with a second argument.

Tools:

* TestWebKitAPI/Tests/WTF/WeakPtr.cpp: Removed use of the DID_CREATE_WEAK_PTR_IMPL
and WILL_DESTROY_WEAK_PTR_IMPL, which didn't work consistently because of the
one-definition rule. Instead, this file parameterizes the WeakPtr family of
class templates with a counter class. To avoid having to rewrite test code, used
alias templates to insert these class template into the TestWebKitAPI namespace,
hiding the class templates from the WTF namespace.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261818 268f45cc-cd09-0410-ab3c-d52691b4dbfc