DataTransfer.files contains multiple files when pasting a single image with multiple representations https://bugs.webkit.org/show_bug.cgi?id=212245 <rdar://problem/60240436>
Reviewed by Tim Horton.

Source/WebCore:

When pasting or dropping a single image that is backed by multiple representations in NSPasteboard (or
UIPasteboard), we currently report more than one `File` to the page via `DataTransfer.files`. This is because
`Pasteboard::read(PasteboardFileReader&)`, which is responsible for converting the contents of the pasteboard
into a list of files, currently iterates over every pasteboard type and adds each of them as a file. This is
wrong when an item has multiple type representations.

To differentiate the case where a single item has multiple representations from the case where it has multiple
pasteboard items, we use `allPasteboardItemInfo()` instead to grab a per-item list of types from the pasteboard
on Cocoa platforms, and only create at most 1 file per item using the highest fidelity type that contains data.

Test: PasteImage.PasteImageWithMultipleRepresentations

* platform/cocoa/PasteboardCocoa.mm:
(WebCore::Pasteboard::read):

Tools:

* DumpRenderTree/mac/DumpRenderTreePasteboard.mm:
(-[LocalPasteboard _clearContentsWithoutUpdatingChangeCount]):
(-[LocalPasteboard _addTypesWithoutUpdatingChangeCount:owner:]):
(-[LocalPasteboard writeObjects:]):
(-[LocalPasteboard pasteboardItems]):

Adjust DumpRenderTree's LocalPasteboard so that it lazily populates the pasteboard when constructing
NSPasteboardItems. To do this, we need to make a few adjustments:

1.      When reifying NSPasteboardItems from LocalPasteboard, ask the owner (WebHTMLView) to provide pasteboard
        data for each pasteboard type that was promised by WebKit, but was not eagerly written to the pasteboard.

2.      Cache pasteboard items that were created, so that we don't repeatedly ask WebHTMLView to provide
        pasteboard data. WebHTMLView doesn't currently support this, and suffers from a bug where TIFF data may
        only be provided once. This was fixed for WebKit2, but not for WebKit1.

3.      Maintain a separate hash list of original pasteboard types (which may not be UTIs) that were handed to
        LocalPasteboard by WebKit. We use these original types in step (1).

* TestWebKitAPI/Tests/WebKitCocoa/PasteImage.mm:

Add a new API test to verify that one file is exposed via the DataTranfer when the pasteboard contains a single
image with two image representations, but two files are exposed when the pasteboard contains two images, each
with a single representation.

(writeImageDataToPasteboard):

Overload this helper method with two additional variants: one that takes a dictionary of pasteboard types to
data, and another that takes an array of dictionaries, each representing a single item's types and data.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262047 268f45cc-cd09-0410-ab3c-d52691b4dbfc