[Cocoa] fast/text/font-collection.html is failing https://bugs.webkit.org/show_bug.cgi?id=212172 <rdar://problem/63466187>
Unreviewed.

The test is failing because the collection font file placed one of its table
directories after the table data. This is a valid file, so this patch doesn't
delete the test, but it also duplicates the test with a new collection font
file that doesn't have this construction. The duplicate test should pass on
all OSes.

* fast/text/font-collection-2-expected.html: Copied from LayoutTests/fast/text/font-collection.html.
* fast/text/font-collection-2.html: Added. Add some extra characters just to
  make sure things are working properly.
* fast/text/font-collection-expected.html:
* fast/text/font-collection.html: Ditto.
* fast/text/resources/collection2.ttc: Added.
* platform/ios/TestExpectations:
* platform/mac/TestExpectations:
* platform/win/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262060 268f45cc-cd09-0410-ab3c-d52691b4dbfc