Updates to the isolated tree must happen before posting notifications to clients. https://bugs.webkit.org/show_bug.cgi?id=212266
Reviewed by Chris Fleizach.

Multiple tests.

In AXObjectCache::notificationPostTimerFired we were updating the
isolated tree after the notifications were posted to the platform
clients. This caused that in some cases when the client requested info
as the result of those notifications, the isolated tree was out-of-date.
In this patch updateIsolatedTree is called before notifying platform clients.

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::notificationPostTimerFired):
(WebCore::AXObjectCache::postNotification):
(WebCore::AXObjectCache::postTextStateChangeNotification):
(WebCore::AXObjectCache::updateIsolatedTree):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262073 268f45cc-cd09-0410-ab3c-d52691b4dbfc