REGRESSION (r261812): editing/async-clipboard/clipboard-item-get-type-basic.html is flaky https://bugs.webkit.org/show_bug.cgi?id=212281 <rdar://problem/63554912>
Reviewed by Tim Horton.

Source/WebKit:

Declare `-delegateSupportsImagePaste` on UIKeyboardImpl.

* Platform/spi/ios/UIKitSPI.h:

Source/WTF:

Add a new `HAVE()` define.

* wtf/PlatformHave.h:

Tools:

On some shipping versions of iOS, returning `NO` from `-supportsImagePaste` still results in UIKit keyboard
code pinning temporary items to the general pasteboard, which increments the change count of the pasteboard.
If this happens in the middle of an attempt to read from the pasteboard, we end up falsely denying access to the
contents of the pasteboard, since we believe that the contents of the pasteboard have changed.

This has the potential to affect any test that attempts to read from the pasteboard on iOS, though the titular
layout test seems to trigger the bug more frequently than other tests.

This item pinning was added in support of being able to insert Memojis from the software keyboard, and works by
pretending to copy a temporary PNG image, asking the delegate whether it `-canPerformAction:withSender:`, and
then restoring the items previously on the pasteboard. To work around this in the test runner, we can simply
swizzle out `-[UIKeyboardImpl delegateSupportsImagePaste]` to always return `NO`, which has the same effect as
disabling the Memoji keyboard.

* WebKitTestRunner/TestController.h:
* WebKitTestRunner/ios/TestControllerIOS.mm:
(overrideKeyboardDelegateSupportsImagePaste):
(WTR::TestController::platformResetStateToConsistentValues):

LayoutTests:

Remove the failing test expectation.

* platform/ios-simulator-wk2/TestExpectations:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262283 268f45cc-cd09-0410-ab3c-d52691b4dbfc