[JSC] JSBigInt::rightTrim can miss |this| pointer and leads to incorrect GC collection https://bugs.webkit.org/show_bug.cgi?id=212601
Reviewed by Saam Barati.

JSTests:

* stress/bigint-should-not-be-collected-while-creating.js: Added.
(foo.let.increment.10000n.bar):
(foo):

Source/JavaScriptCore:

This is pretty rare case. But in some optimization level, JSBigInt::rightTrim could store |this| + offset pointer into the stack instead of |this|
and make conservative GC think that |this| JSBigInt is unreachable. We put ensureStillAliveHere(this) to ensure that this is alive.

* runtime/JSBigInt.cpp:
(JSC::JSBigInt::rightTrim):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262392 268f45cc-cd09-0410-ab3c-d52691b4dbfc