JSTests: Bytecode liveness should be aware of checkpoints https://bugs.webkit.org/show_bug.cgi?id=213106 <rdar://problem/63416838>
Reviewed by Keith Miller.

* stress/iterator-open-checkpoint-liveness.js: Added.
(f):
(next):
(iterator):
(catch):

Source/JavaScriptCore:
op_iterator_open getNext checkpoint needs to declare it uses m_iterator
https://bugs.webkit.org/show_bug.cgi?id=213106
<rdar://problem/63416838>

 Reviewed by Keith Miller.

Currently, we have no way of specifying that a checkpoint uses an operand defined at an earlier
point in the same bytecode, which is the case for op_iterator_open: we assume that it will have
already allocated the iterator and stored it in m_iterator by the time we get to the getNext
checkpoint. In order to support that, we change tmpLivenessForCheckpoint to livenessForCheckpoint
and allow it to also declare the use of the operands defined within the bytecode.


* bytecode/BytecodeLivenessAnalysis.cpp:
(JSC::livenessForCheckpoint):
(JSC::tmpLivenessForCheckpoint): Deleted.
* bytecode/BytecodeLivenessAnalysis.h:
* bytecode/FullBytecodeLiveness.h:
* dfg/DFGForAllKills.h:
(JSC::DFG::forAllKilledOperands):
* dfg/DFGGraph.cpp:
(JSC::DFG::Graph::isLiveInBytecode):
* dfg/DFGGraph.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263071 268f45cc-cd09-0410-ab3c-d52691b4dbfc