$vm.haveABadTime/isHavingABadTime should work with non-globalObject parameters https://bugs.webkit.org/show_bug.cgi?id=213304
Reviewed by Mark Lam.

JSTests:

* stress/dollarVM-have-a-bad-time-no-params.js: Added.
* stress/dollarVM-have-a-bad-time-works-for-non-global-object-params.js: Added.

Source/JavaScriptCore:

Previously, $vm.haveABadTime would crash if passed a
non-globalObject object as the first parameter because it was
missing a `return` in front the error handling case. This patch
resolves that issue but also extends the semantics of
haveABadTime/isHavingABadTime to either use the global object of
the first parameter even if it's not a JSGlobalObject. If no
argument is passed, haveABadTime/isHavingABadTime instead use the
global object of the callee.

* tools/JSDollarVM.cpp:
(JSC::functionHaveABadTime):
(JSC::functionIsHavingABadTime):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263174 268f45cc-cd09-0410-ab3c-d52691b4dbfc