[JSC] Freeze JSBigInt when setting it as a constant in AI https://bugs.webkit.org/show_bug.cgi?id=213310 <rdar://problem/64450410>
Reviewed by Mark Lam.

JSTests:

* stress/bigint-constant-fold-to-numeric.js: Added.
(foo):

Source/JavaScriptCore:

JSCells should be explicitly frozen via DFG::Graph::freeze or DFG::Graph::freezeStrong. And heap JSBigInt is JSCell.
We should freeze it before setting it as a parameter of setConstant in AI. We use DFG::Graph::freeze since we know
that this is coming from somewhere in DFG graph: this ToNumeric node itself is not newly producing this JSBigInt.

* dfg/DFGAbstractInterpreterInlines.h:
(JSC::DFG::AbstractInterpreter<AbstractStateType>::executeEffects):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263180 268f45cc-cd09-0410-ab3c-d52691b4dbfc