[webkitpy][WPE] Default to headless if _display_server is xfvb https://bugs.webkit.org/show_bug.cgi?id=213327
Reviewed by Philippe Normand.

Xvfb is used as the default display server for some scripts, and the
equivalent for WPE is the headless one.

WTR and the API tests already create only HeadlessViewBackends, but the
WebDriver tests create them through MiniBrowser, which requires the
'--headless' parameter.

* Scripts/run-webdriver-tests: Do not override _display_server
* Scripts/webkitpy/port/base.py:
(Port.__init__): Define _display_server for all ports to avoid
eventually needing to override it.
* Scripts/webkitpy/port/wpe.py:
(WPEPort.__init__): Default to 'headless' if _display_server is 'xvfb'
* Scripts/webkitpy/w3c/wpt_runner.py:
(main): Dot not override _display_server


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263261 268f45cc-cd09-0410-ab3c-d52691b4dbfc