com.apple.WebKit.Networking crash: suspended with locked system files (observations.db) https://bugs.webkit.org/show_bug.cgi?id=213391 <rdar://problem/64494167>
Reviewed by Chris Dumez.

ITP is the only database that attempts to close when
NetworkProcess::didClose() is called. This causes the network process
to sometimes crash if the database is still closing when checking for
locked system files. Instead, we should only flush the
memory store to disk when calling NetworkProcess::didClose (in the
memory store case), and only close the database when the network session is destroyed.

* NetworkProcess/Classifier/WebResourceLoadStatisticsStore.cpp:
(WebKit::WebResourceLoadStatisticsStore::didDestroyNetworkSession):
(WebKit::WebResourceLoadStatisticsStore::destroyResourceLoadStatisticsStore):
(WebKit::WebResourceLoadStatisticsStore::flushAndDestroyPersistentStore):
* NetworkProcess/Classifier/WebResourceLoadStatisticsStore.h:
Split into two functions, one to flush and destroy the persistent
store, and one to destroy the memory or database store.

* NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::didClose):
Only flush the persistent store here, as described above.

* NetworkProcess/NetworkSession.cpp:
(WebKit::NetworkSession::flushAndDestroyPersistentStore):
* NetworkProcess/NetworkSession.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263298 268f45cc-cd09-0410-ab3c-d52691b4dbfc