REGRESSION (r263253): Search field results and cancel buttons have their own focus rings https://bugs.webkit.org/show_bug.cgi?id=213413 <rdar://problem/64548419>
Reviewed by Tim Horton.

Source/WebCore:

After r263253, `paintCellAndSetFocusedElementNeedsRepaintIfNecessary` is used when painting the buttons in a
search field's shadow root. However, the renderer that is passed in (which is used to determine whether we
should additionally draw focus rings) is the input element's renderer rather than the renderers of the results
and cancel buttons themselves. This means that when the search field is focused, we will draw focus rings around
each of the buttons in the shadow root as well.

Address this by using `box` (the buttons' RenderBoxes) instead.

Test: fast/forms/search-field-buttons-do-not-have-focus-rings.html

* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::paintSearchFieldCancelButton):
(WebCore::RenderThemeMac::paintSearchFieldResultsButton):

LayoutTests:

Add a ref test to verify that the search and cancel buttons don't draw individual focus rings. This test works
by verifying that no focus ring around either the cancel or results button bleeds outside of the bounds of a div
that obscures the buttons.

* fast/forms/search-field-buttons-do-not-have-focus-rings-expected.html: Added.
* fast/forms/search-field-buttons-do-not-have-focus-rings.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263303 268f45cc-cd09-0410-ab3c-d52691b4dbfc