[LFC][TFC] Adjust table content vertical position to match vertical-align https://bugs.webkit.org/show_bug.cgi?id=213692
Reviewed by Antti Koivisto.

Source/WebCore:

Child boxes (and runs) are always in the coordinate system of the containing block's border box.
The content box (where the child content lives) is inside the padding box, which is inside the border box.
In order to compute the child box top/left position, we need to know both the padding and the border offsets.

<div style="border: 2px solid green; padding-top: 10px;"><div></div></div>
The inner div's top left position is at [12px, 2px] (let's ignore margin collapsing for now).

Normally by the time we start positioning the child content, we already have computed borders and paddings for the containing block.
This is different with table cells where the final padding offset depends on the content height as we use
the padding box to vertically align the table cell content.

Let's adjust the child boxes and runs to match the new content box position.

Test: fast/layoutformattingcontext/table-cell-vertical-alignment-simple.html

* layout/displaytree/DisplayLineBox.h:
(WebCore::Display::LineBox::moveVertically):
* layout/displaytree/DisplayRun.h:
(WebCore::Display::Run::moveVertically):
* layout/tableformatting/TableFormattingContext.cpp:
(WebCore::Layout::TableFormattingContext::setUsedGeometryForCells):

LayoutTests:

* fast/layoutformattingcontext/table-cell-vertical-alignment-simple-expected.html: Added.
* fast/layoutformattingcontext/table-cell-vertical-alignment-simple.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263640 268f45cc-cd09-0410-ab3c-d52691b4dbfc