[Media Controls] Tracks panel can show text tracks with mode "hidden" as selected https://bugs.webkit.org/show_bug.cgi?id=213839 <rdar://problem/57989325>
Reviewed by Jer Noble.

Source/WebCore:

We only considered text tracks that had their "mode" set to "disabled" as tracks that weren't selected,
but there is also the "hidden" mode which should share the same UI state. We now check for "mode" to
be set to "showing" to consider a text track as selected in the tracks panel.

Test: media/modern-media-controls/tracks-support/tracks-support-hidden-tracks.html

* Modules/modern-media-controls/media/tracks-support.js:
(TracksSupport.prototype.tracksPanelIsTrackInSectionSelected):

LayoutTests:

Add a new test that dumps the selection state of tracks in the tracks panel when all of media.textTracks
have been set to mode = "hidden".

* media/modern-media-controls/tracks-support/tracks-support-hidden-tracks-expected.txt: Added.
* media/modern-media-controls/tracks-support/tracks-support-hidden-tracks.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263802 268f45cc-cd09-0410-ab3c-d52691b4dbfc