[CMake] Add WOFF2 targets https://bugs.webkit.org/show_bug.cgi?id=213865
Reviewed by Fujii Hironori.

.:

Rework the Find module for WOFF2 to provide targets for libwoff2common and
libwoff2dec. Previously only libwoff2dec was specified in the find module. This
is because pkg-config works on that platform and because shared libraries are
supported. Neither is true for Windows so the module was expanded in anticipation
of support for WinCairo.

* Source/cmake/FindWOFF2.cmake: Added.
* Source/cmake/FindWOFF2Dec.cmake: Removed.
* Source/cmake/OptionsGTK.cmake:
* Source/cmake/OptionsWPE.cmake:

Source/WebCore:

Use the WOFF:: targets when USE_WOFF2 is set.

* CMakeLists.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263829 268f45cc-cd09-0410-ab3c-d52691b4dbfc